+++ src/tests/unit/test_tree_functionality/mock_data_for_tree_tests/mock_blueprints_for_tree_tests/ChestWithOptionalBoxInside.blueprint.json
